refactoring(calculator): 문자열을 int[]로 변환하는 로직을 Parser class로 이동
