+++ src/main/java/calculator/Parser.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import java.util.Arrays;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -7,13 +8,21 @@ public class Parser {
     private static final String DEFAULT_DELIMITER_REGEX = "[,:]";
     private static final Pattern CUSTOM_DELIMITER_REGEX = Pattern.compile("//(.*)\n(.*)");
 
-    public String[] parse(String text) {
+    public int[] parseToIntArray(String text) {
         Matcher m = CUSTOM_DELIMITER_REGEX.matcher(text);
         if (m.find()) {
             String delimiter = m.group(1);
-            return m.group(2).split(delimiter);
+            return convertInt(m.group(2).split(delimiter));
         }
 
-        return text.split(DEFAULT_DELIMITER_REGEX);
+        return convertInt(text.split(DEFAULT_DELIMITER_REGEX));
+    }
+
+    private int[] convertInt(String[] stringNumbers) {
+        try {
+            return Arrays.stream(stringNumbers).mapToInt(Integer::parseInt).toArray();
+        } catch (NumberFormatException e) {
+            throw new NumberFormatException("``stringNumbers` is must be number format");
+        }
     }
 }
