+++ src/main/java/calculator/README.md
@@ -14,12 +14,12 @@
 * 구분자로 연결된 0, 자연수들을 입력했을때 그들의 합을 계산하는 기능을 구현한다.
 * 쉼표(,) 또는 콜론(:)을 구분자로 가지는 문자열을 전달하는 경우 구분자를 기준으로 분리한 각 숫자의 합을 반환 (예: “” => 0, "1,2" => 3, "1,2,3" => 6, “1,2:3” => 6)
   * public StringAddCalculator::sum(String input) String
-  * private StringAddCalculator::split(String input, String delimiter) String[]
   * 쉼표만 있는 경우, 콜론만 있는 경우, 같이 있는 경우, 구분자가 없는 경우 테스트
 * 앞의 기본 구분자(쉼표, 콜론)외에 커스텀 구분자를 지정할 수 있다. 커스텀 구분자는 문자열 앞부분의 “//”와 “\n” 사이에 위치하는 문자를 커스텀 구분자로 사용한다. 예를 들어 “//;\n1;2;3”과 같이 값을 입력할 경우 커스텀 구분자는 세미콜론(;)이며, 결과 값은 6이 반환되어야 한다.
   * Parser class
      * 기본 구분자, 커스텀 구분자의 유무를 파악하고 문자열을 String[] 형태로 파싱한다.
-     * public parse(String text) String[]
+     * public parseToIntArray(String text) int[]
+     * private convertInt(String[] stringNumbers) int[]
   * 커스텀 구분자가 있는 경우, 없는 경우 테스트
   * 커스텀 구분가의 형식이 틀린 경우 테스트
 * 문자열 계산기에 숫자 이외의 값 또는 음수를 전달하는 경우 RuntimeException 예외를 throw한다.
