+++ src/main/java/calculator/StringAddCalculator.java
@@ -11,18 +11,8 @@ public class StringAddCalculator {
     }
 
     public int sum(String input) {
-        String[] stringNumbers = parser.parse(input);
-
-        int[] numbers = convertInt(stringNumbers);
+        int[] numbers = parser.parseToIntArray(input);
 
         return Arrays.stream(numbers).reduce(Integer::sum).getAsInt();
-    }
-
-    private int[] convertInt(String[] stringNumbers) {
-        try {
-            return Arrays.stream(stringNumbers).mapToInt(Integer::parseInt).toArray();
-        } catch (NumberFormatException e) {
-            throw new NumberFormatException("``stringNumbers` is must be number format");
-        }
     }
 }
